requirements.txt
requests
feedparser
newspaper3k
openpyxl